pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml_object_detector"
version = "0.1.0"
dependencies = [
  "aiofiles",
  "python-magic"
]

[project.scripts]
ml-etl = "ml_object_detector.cli.run_etl:main"
ml-pipeline = "ml_object_detector.cli.run_pipeline:main"
ml-api = "ml_object_detector.api.run_api:main"

[project.optional-dependencies]
dev = [
  "pre-commit>=3.7",
  "nbconvert>=7.16",
  "nbdime>=3.2",
]
api = [
  "fastapi==0.115.14",
  "uvicorn[standard]==0.35.0",
  "python-multipart==0.0.20"
]

[tool.setuptools]
package-dir = {"" = "src"}
